Fix missing dependency to packaging
When installing the package from pypi with pip not all necessary
dependencies are installed.
In setup.py the dependency to 'packaging' is missing.
setup.py now reads the requirements.txt file and adds the dependencies
to the install_requires field.
That makes requirements.txt the single point were dependencies are
stored.